% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_skewness.R
\name{detect_skewness}
\alias{detect_skewness}
\title{Detect Skewness}
\usage{
detect_skewness(sample)
}
\arguments{
\item{A}{vector or list}
}
\value{
A character string
}
\description{
Detect Skewness
}
\examples{
detect_skewness(c(1, 2, 3, -5, 5, 5, 6, 8, 10))
}
